Add support for mksh
`frum init` produces an error about not being able to infer the shell
when ran under mksh.

From looking at frum's source and documentation, the `__frumcd`
shell function and the `eval` and `frum init` commands for bash all
seem like they will work just fine with mksh. I added mksh alongside
bash and sh and rebuilt frum and everything is working well for me.

I do not currently have access to other ksh derivatives to test with,
and I am not sure of the current state of ksh derivative support in the
"shell" crate. But this simple change has me up and running and enjoying
frum.